Move and redo custom module.

+++ drupal/core/modules/path/src/Plugin/Field/FieldType/PathItem.php
@@ -55,29 +55,25 @@ class PathItem extends FieldItemBase {
   /**
    * {@inheritdoc}
    */
-  public function insert() {
-    if ($this->alias) {
-      $entity = $this->getEntity();
-
-      if ($path = \Drupal::service('path.alias_storage')->save($entity->getSystemPath(), $this->alias, $this->getLangcode())) {
-        $this->pid = $path['pid'];
+  public function postSave($update) {
+    if (!$update) {
+      if ($this->alias) {
+        $entity = $this->getEntity();
+        if ($path = \Drupal::service('path.alias_storage')->save('/' . $entity->urlInfo()->getInternalPath(), $this->alias, $this->getLangcode())) {
+          $this->pid = $path['pid'];
+        }
       }
     }
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function update() {
-    // Delete old alias if user erased it.
-    if ($this->pid && !$this->alias) {
-      \Drupal::service('path.alias_storage')->delete(array('pid' => $this->pid));
-    }
-    // Only save a non-empty alias.
-    elseif ($this->alias) {
-      $entity = $this->getEntity();
-
-      \Drupal::service('path.alias_storage')->save($entity->getSystemPath(), $this->alias, $this->getLangcode(), $this->pid);
+    else {
+      // Delete old alias if user erased it.
+      if ($this->pid && !$this->alias) {
+        \Drupal::service('path.alias_storage')->delete(array('pid' => $this->pid));
+      }
+      // Only save a non-empty alias.
+      elseif ($this->alias) {
+        $entity = $this->getEntity();
+        \Drupal::service('path.alias_storage')->save('/' . $entity->urlInfo()->getInternalPath(), $this->alias, $this->getLangcode(), $this->pid);
+      }
     }
   }
 
@@ -87,7 +83,7 @@ class PathItem extends FieldItemBase {
   public function delete() {
     // Delete all aliases associated with this entity.
     $entity = $this->getEntity();
-    \Drupal::service('path.alias_storage')->delete(array('source' => $entity->getSystemPath()));
+    \Drupal::service('path.alias_storage')->delete(array('source' => '/' . $entity->urlInfo()->getInternalPath()));
   }
 
   /**
